Create snippet for variables in ctor
When creating variables in __init__, I tend to have the same pattern.
This is often done for me by the class snippet, but if I terminate that
snippet too soon, then I want to be able to do this.

Uses Java's snippet trigger for the same idiom: cv (constructor
variable).