cs: Figure out project and module for namespaces
Instead of using the folder name (which is terrible when you're dumping
code into Scripts/), figure out the project (based on Unity filepath
standard) and only use module if it's not 'Scripts'.

Also match company name for case. Not sure of a better way to handle
that. I think we usually want MixedCase, but if company is lower then it
must be some exception.