lua: Extract base class call to new snippet
I don't have the time to convert it to just copying the base args, so
make it another function-creating snippet for now.

Fix arg auto copying and indent. The snip.c check prevented the args
from updating. I think nothing is cyclic so it's not necessary. We only
had to snip.shift because we were inserting our own newlines. Making the
snippet start after an indent solves the indent.