lua: Look at above functions to guess signature
Instead of looking from the top of the file, first try looking from the
current line up to see if an existing function or class definition
exist.

We want to define a function for the same class as the above function or
the closest class.

Add some tests since that makes snippet dev so much easier. I use pytest
now, but these tests could be run manually too.